requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Pillow==9.0.0
python-dotenv==0.19.0
